Moved tracker classes to tracker package
