+++ fortify-issue-manager/src/main/java/com/microfocus/security/automation/fortify/issue/tracker/OctaneBugTrackerSettings.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.microfocus.security.automation.fortify.issue.manager;
+package com.microfocus.security.automation.fortify.issue.tracker;
 
 import java.util.Map;
